Refactor: 수정 api
