+++ app/src/main/java/com/example/mumuk/data/model/allergy/AllergyOption.kt
@@ -1,0 +1,6 @@
+package com.example.mumuk.data.model.allergy
+
+data class AllergyOption(
+    val id: Int,
+    val allergyType: String
+)
