+++ app/src/main/java/com/example/mumuk/data/model/allergy/ToggleAllergyRequest.kt
@@ -1,0 +1,5 @@
+package com.example.mumuk.data.model.allergy
+
+data class ToggleAllergyRequest(
+    val allergyTypeList: List<String>
+)
